Refactor dashboard endpoints and options
