+++ src/Immediate.Jobs.Dashboard/GetDashboardJobTelemetryLinks.cs
@@ -1,0 +1,40 @@
+using Immediate.Apis.Shared;
+using Immediate.Handlers.Shared;
+using Immediate.Jobs.Shared.Interfaces;
+using Immediate.Validations.Shared;
+using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.Extensions.Options;
+
+#pragma warning disable CA1812 // Request types and route groups are activated by generated endpoints.
+
+namespace Immediate.Jobs.Dashboard;
+
+[Handler]
+[MapGet("jobs/{jobId}/telemetry-links")]
+[MapGroup<DashboardApi>]
+internal static partial class GetDashboardJobTelemetryLinks
+{
+	[Validate]
+	internal sealed partial record Query : IValidationTarget<Query>
+	{
+		[NotEmpty]
+		public required string JobId { get; init; }
+	}
+
+	internal static Results<JsonHttpResult<IReadOnlyList<JobTelemetryLink>>, NotFound> TransformResult(
+		IReadOnlyList<JobTelemetryLink>? result
+	) => DashboardApiEndpointOperations.TransformTelemetryLinksResult(result);
+
+	private static ValueTask<IReadOnlyList<JobTelemetryLink>?> HandleAsync(
+		Query query,
+		IJobMonitor monitor,
+		IOptions<ImmediateJobsDashboardOptions> options,
+		CancellationToken cancellationToken
+	) => DashboardApiEndpointOperations.GetJobTelemetryLinksAsync(
+		query.JobId,
+		executionNumber: null,
+		monitor,
+		options.Value,
+		cancellationToken
+	);
+}
