+++ src/Immediate.Jobs.Dashboard/ImmediateJobsDashboardEndpointRouteBuilderExtensions.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 
 namespace Immediate.Jobs.Dashboard;
 
@@ -11,22 +12,18 @@ public static class ImmediateJobsDashboardEndpointRouteBuilderExtensions
 {
 	/// <summary>Maps the dashboard at <c>/jobs</c>.</summary>
 	/// <param name="endpoints">The endpoint route builder to add the dashboard to.</param>
-	/// <param name="configure">An optional callback that configures the dashboard.</param>
 	/// <returns>The route group containing the dashboard endpoints.</returns>
 	public static RouteGroupBuilder MapImmediateJobsDashboard(
-		this IEndpointRouteBuilder endpoints,
-		Action<ImmediateJobsDashboardOptions>? configure = null
-	) => endpoints.MapImmediateJobsDashboard("/jobs", configure);
+		this IEndpointRouteBuilder endpoints
+	) => endpoints.MapImmediateJobsDashboard("/jobs");
 
 	/// <summary>Maps the dashboard and its API under <paramref name="prefix"/>.</summary>
 	/// <param name="endpoints">The endpoint route builder to add the dashboard to.</param>
 	/// <param name="prefix">The URL path prefix under which to map the dashboard.</param>
-	/// <param name="configure">An optional callback that configures the dashboard.</param>
 	/// <returns>The route group containing the dashboard endpoints.</returns>
 	public static RouteGroupBuilder MapImmediateJobsDashboard(
 		this IEndpointRouteBuilder endpoints,
-		string prefix,
-		Action<ImmediateJobsDashboardOptions>? configure = null
+		string prefix
 	)
 	{
 		ArgumentNullException.ThrowIfNull(endpoints);
@@ -38,12 +35,10 @@ public static class ImmediateJobsDashboardEndpointRouteBuilderExtensions
 		if (prefix.Length == 0)
 			throw new ArgumentException("The dashboard must be mapped below the application root.", nameof(prefix));
 
-		var options = endpoints.ServiceProvider.GetService<ImmediateJobsDashboardOptions>()
+		var options = endpoints.ServiceProvider.GetService<IOptions<ImmediateJobsDashboardOptions>>()?.Value
 			?? throw new InvalidOperationException(
 				"Dashboard services are not registered. Call services.AddImmediateJobsDashboard() before building the application."
 			);
-		configure?.Invoke(options);
-		options.Validate();
 
 		var group = endpoints.MapGroup(prefix).WithTags("Immediate.Jobs Dashboard");
 		if (options.AuthorizationPolicy is { } policy)
