+++ src/Immediate.Jobs.Redis/Immediate.Jobs.Redis.csproj
@@ -21,6 +21,7 @@
   </ItemGroup>
 
   <ItemGroup>
+	<PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" />
 	<PackageReference Include="StackExchange.Redis" />
   </ItemGroup>
